requirements.txt
APScheduler==3.2.0
blessed==1.14.1
click==6.6
coverage==4.2
fitbit==0.2.4
Flask==0.11.1
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.1
funcsigs==1.0.2
futures==3.0.5
gunicorn==19.6.0
honcho==0.7.1
importlib==1.0.3
itsdangerous==0.24
Jinja2==2.8
keyring==8.7
keyrings.alt==1.1.1
logging==0.4.9.6
lxml==3.6.1
MarkupSafe==0.23
measurement==1.8.0
mimerender==0.6.0
myfitnesspal==1.8.6
MySQL-python==1.2.5
nose==1.3.7
oauthlib==1.1.2
psycopg2==2.6.2
python-dateutil==2.4.2
python-mimeparse==1.5.2
pytodoist==2.0.5
pytz==2016.6.1
requests==2.11.0
requests-oauth==0.4.1
requests-oauthlib==0.6.2
six==1.10.0
SQLAlchemy==1.0.14
SQLAlchemy-Utils==0.32.9
sympy==0.7.6.1
tweepy==3.5.0
tzlocal==1.2.2
wcwidth==0.1.7
Werkzeug==0.11.10
withings==0.3
wunderpy2==0.1.4
